tests/requirements.txt
scratchcommunication
python-dotenv